refactor and it works on Windows.
